Solves issue #279
